requirements.txt
amqp==5.3.1
asgiref==3.9.1
billiard==4.2.1
celery==5.5.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
coverage==7.10.4
Django==5.2.5
django-health-check==3.20.0
django-ipware==7.0.1
django-prometheus==2.4.1
django-redis==6.0.0
django-structlog==9.1.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.10
inflection==0.5.1
kombu==5.5.4
packaging==25.0
prometheus_client==0.22.1
prompt_toolkit==3.0.51
psutil==7.0.0
psycopg==3.2.9
psycopg-pool==3.2.6
psycopg2==2.9.10
psycopg2-binary==2.9.10
PyJWT==1.7.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-ipware==3.0.0
pytz==2025.2
PyYAML==6.0.2
redis==6.4.0
six==1.17.0
sqlparse==0.5.3
structlog==25.4.0
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
vine==5.1.0
wcwidth==0.2.13
